% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebirdclean.R
\name{ebirdclean}
\alias{ebirdclean}
\title{Clean data frame provided by eBird}
\usage{
ebirdclean(x = "MyEBirdData.csv")
}
\arguments{
\item{x}{CSV file provided by eBird. Defaults to the name provided by eBird: "MyEBirdData.csv".}
}
\value{
An updated data frame, with the following changes:
  - Separates the Date column into two new columns, Year and Month.
  - Extracts country information from State.Province and converts into full name (using the countrycode package)
  - Removes text in parentheses in Common.name, which is used to denote subspecies or groups.
  - Removes all rows of species labelled as hybrids, domestic types, slashes (uncertain records between two species), or spuhs (Genus sp.).
}
\description{
Clean data frame provided by eBird
}
\details{
Uses \code{read.csv} to read the .csv file, with its only argument being \code{stringsAsFactors = FALSE}.
}
\examples{
\dontrun{
mylist <- ebirdclean("MyEBirdData.csv")
mylist <- ebirdclean()}
}
\author{
Sebastian Pardo \email{sebpardo@gmail.com}
}

