% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myebirds.R
\name{myebirds}
\alias{myebirds}
\title{Calculate species totals gouped by year, country, and/or month.}
\usage{
myebirds(mydata, years = 1900:format(Sys.Date(), "\%Y"),
  grouping = c("Year", "Country", "Month"), wide = FALSE)
}
\arguments{
\item{mydata}{Data frame provided by \code{ebirdclean}.}

\item{years}{Range of years to calculate across, default is between 1900 and current year.}

\item{grouping}{Character vector specifying how the data should be grouped for counting.
Must be composed of "Year", "Country", and/or "Month". This vector is passed on directly to \code{group_by_}.}

\item{wide}{String specifying whether different values should be returned in wide format and how.
Must be one of the strings included in \code{grouping}. Defaults to FALSE.}
}
\value{
A data frame containing total counts divided into specified groups. If \code{wide = FALSE}
  then it returns a combination of the following columns, depending on grouping specified:

"Year" Year

"Country" Country using two letter codes.

"Month" Month, using full month name (from month.name()).

"n" Total species count for the specified Year, Country, and Month.

If in wide format, then the first column(s) consist of the values in grouping that
are not equal to wide, while the remaining columns are unique values of the argument specified in wide.
}
\description{
\code{myebirds} calculates species totals based on data dumped by eBird
(Download page: \url{http://ebird.org/ebird/downloadMyData}) after it has been cleaned with \code{ebirdclean}.
}
\examples{
\dontrun{
mylist <- ebirdclean() # CSV must be in working directory
myebirds(mylist, grouping = c("Year"))
myebirds(mylist, grouping = c("Country", "Month"))
myebirds(mylist, grouping = c("Year", "Country", "Month"))
myebirds(mylist, grouping = c("Year", "Country", "Month"), wide = "Month")
myebirds(mylist, grouping = c("Year", "Country", "Month"), wide = "Country")
}
}
\author{
Sebastian Pardo \email{sebpardo@gmail.com}
}

