% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mypd.R
\name{mypd}
\alias{mypd}
\title{Total phylogenetic distance of species in your eBird lists.}
\usage{
mypd(mydata)
}
\arguments{
\item{mydata}{Data frame created with \code{\link{ebirdclean}}.}
}
\value{
A tibble containing the mean ("mean_pd"), median ("median_pd"), and standard
 deviation ("sd_pd") vaules; units are in million years.
}
\description{
\code{mypd} provides the evolutionary time encompassed by your eBird checklist
based on Jetz et al. (2014).
}
\details{
Phylogenetic distance (PD) is simply the sum of all branch lengths of a tree,
and represents the total evolutionary time covered by the species in the tree.
Values returned by \code{mypd} are in million years of evolution, and are based
in a subset of 20 trees from Jetz et al. (2014) using the full taxonomy and
Hackett "structure". See \code{raw-data/load-bird-trees.R} for more info on
how the trees were extracted from the original source.#'
}
\examples{
\dontrun{
mylist <- ebirdclean("MyEBirdData.csv")
mypd(mylist)

# To look at PD values by country (using dplyr):
mybirds \%>\%
  group_by(Country) \%>\%
  do(mypd(.))}

}
\author{
Sebastian Pardo \email{sebpardo@gmail.com}
}
\references{
Jetz, W., Thomas, G. H., Joy, J. B., Redding, D. W.,
  Hartmann, K. & Mooers, A. O. (2014) Global Distribution and Conservation
  of Evolutionary Distinctness in Birds. Current Biology 24, 919–930.
  \url{http://www.cell.com/current-biology/fulltext/S0960-9822(14)00270-X}
}

